+++ NEWS.md
@@ -4,6 +4,7 @@
 * Added no coverage blocks to `ageKeyPlot()`, `GompertzModels()`, `hist.formula()`, `histFromSum()`, `logisticModels()`, `lwCompPreds()`, `plot.agebias()`, `plot.CatchCurve()`, `plot.ChapmanRobson()`, `plot.Depletion()`, `print.compSlopes()`, `print.compIntercepts()`, `print.metaM()`, `psdPlot()`, `RichardsModels()`, `SchnuteModel()`, `srModels()`, `srStarts()`, `vbModels()`, and `vbStarts()`.
 * `ageKey()`: Removed.  Deprecated since 0.4.24.  Use `alkIndivAge()`.
 * `ageKeyPlot()`: Removed.  Deprecated since 0.4.24.  Use `alkPlot()`.
+* `bcFuns()`: Modified.  Changed `msg=` to `verbose=`.
 * `filterD()`: Modified.  Changed to use `droplevels()` from `base` rather than `drop.levels()` from `gdata`.  Added `except=`.
 * `fitPlot()`: Modified.  Changed the way colors, plotting characters, and line types were handled for most of the models.  Should make their use more flexible.  Started to add tests for error and warning messages.
 * `lenFreqExpand()`: Removed.  Deprecated since 0.4.32.  Use `expandLenFreq()`.